+++ CHANGELOG.md
@@ -4,6 +4,8 @@ LifterLMS Changelog
 v3.23.1 - 2018-08-??
 --------------------
 
++ Remove use of deprecated `LLMS_Lesson->get_children_lessons()` in the `LLMS_Course` model
++ Minor increase to performance in the `LLMS_Course->get_lessons()` method
 + Updated all access plan templates added in 3.23.0 to ensure `ABSPATH` is defined to prevent direct template access
 
 ##### Template Updates